new theme for WAS
Also included a new class to apply momentum scrolling to scrolling overflow elements.
Note that to compile each theme, you need to modify the theme variable in the gulpfile. Also, to use the index.html test view, you need to point the style tag to the appropriate theme. I'm sure both could be automated using a gulp command that pulls in the theme name and appends the correct style sheet to the index HTML.