Updated clib and was theme with .smaller class and font weight classes
Not sure if font weight classes are going to work well. Open sans has an extra-bold 800 weight and source sans pro has a black 900 weight which function the same. Hopefully by including both weights in the class, open sans will default to the 800 weight as there is no 900 weight and source sans pro will default to the 900 weight because it comes after the 800 value in the style rule